Add footer to modal and style footer and scrollbar
